+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Changelog for cfgrib
   See `#294 <https://github.com/ecmwf/cfgrib/pull/294/>`_.
 - fixed support for GRIB files with alternativeRowScanning=1
   See  `#296 <https://github.com/ecmwf/cfgrib/pull/296/>`_.
+- fixed support for missing values
+- See `#313 <https://github.com/ecmwf/cfgrib/issues/313>`_.
 
 
 0.9.10.1 (2022-03-16)